Reorganize the test helpers and use delegate for double timeout waiter
