+++ test/LibraryManager.IntegrationTest/LibmanCompletionTests.cs
@@ -30,7 +30,7 @@ namespace Microsoft.Web.LibraryManager.IntegrationTest
             Editor.Caret.MoveDown(2);
             Editor.KeyboardCommands.Type("\"destination\":");
 
-            LibmanTestsUtility.WaitForCompletionEntries(Editor, expectedCompletionEntries, caseInsensitive: true);
+            CompletionTestsHelper.WaitForCompletionEntries(Editor, expectedCompletionEntries, caseInsensitive: true);
         }
 
         [TestMethod]
@@ -48,7 +48,7 @@ namespace Microsoft.Web.LibraryManager.IntegrationTest
             Editor.Caret.MoveDown(2);
             Editor.KeyboardCommands.Type("\"provider\":");
 
-            LibmanTestsUtility.WaitForCompletionEntries(Editor, expectedCompletionEntries, caseInsensitive: true);
+            CompletionTestsHelper.WaitForCompletionEntries(Editor, expectedCompletionEntries, caseInsensitive: true);
         }
 
         [TestMethod]
@@ -67,7 +67,7 @@ namespace Microsoft.Web.LibraryManager.IntegrationTest
             Editor.KeyboardCommands.Enter();
             Editor.KeyboardCommands.Type("\"defaultProvider\":");
 
-            LibmanTestsUtility.WaitForCompletionEntries(Editor, expectedCompletionEntries, caseInsensitive: true);
+            CompletionTestsHelper.WaitForCompletionEntries(Editor, expectedCompletionEntries, caseInsensitive: true);
         }
 
         [TestMethod]
@@ -84,7 +84,7 @@ namespace Microsoft.Web.LibraryManager.IntegrationTest
             Editor.KeyboardCommands.Enter();
             Editor.KeyboardCommands.Type("\"defaultDestination\":");
 
-            LibmanTestsUtility.WaitForCompletionEntries(Editor, expectedCompletionEntries, caseInsensitive: true);
+            CompletionTestsHelper.WaitForCompletionEntries(Editor, expectedCompletionEntries, caseInsensitive: true);
         }
 
         [TestMethod]
@@ -101,7 +101,7 @@ namespace Microsoft.Web.LibraryManager.IntegrationTest
             Editor.KeyboardCommands.Enter();
 
             Editor.KeyboardCommands.Type("\"library\":");
-            LibmanTestsUtility.WaitForCompletionEntries(Editor, expectedCompletionEntries, caseInsensitive: true, timeout: 5000);
+            CompletionTestsHelper.WaitForCompletionEntries(Editor, expectedCompletionEntries, caseInsensitive: true, timeout: 5000);
         }
 
         [TestMethod]
@@ -116,10 +116,10 @@ namespace Microsoft.Web.LibraryManager.IntegrationTest
             Editor.KeyboardCommands.Enter();
 
             Editor.KeyboardCommands.Type("\"library\":");
-            LibmanTestsUtility.WaitForCompletionEntry(Editor, "jquery", caseInsensitive: true, timeout: 5000);
+            CompletionTestsHelper.WaitForCompletionEntry(Editor, "jquery", caseInsensitive: true, timeout: 5000);
 
             Editor.KeyboardCommands.Type("jquery@");
-            LibmanTestsUtility.WaitForCompletionEntries(Editor, new string[] { }, caseInsensitive: true);
+            CompletionTestsHelper.WaitForCompletionEntries(Editor, new string[] { }, caseInsensitive: true);
         }
 
         [TestMethod]
@@ -137,7 +137,7 @@ namespace Microsoft.Web.LibraryManager.IntegrationTest
             Editor.KeyboardCommands.Enter();
 
             Editor.KeyboardCommands.Type("\"library\":");
-            LibmanTestsUtility.WaitForCompletionEntries(Editor, expectedCompletionEntries, caseInsensitive: true);
+            CompletionTestsHelper.WaitForCompletionEntries(Editor, expectedCompletionEntries, caseInsensitive: true);
         }
 
         [TestMethod]
@@ -156,7 +156,7 @@ namespace Microsoft.Web.LibraryManager.IntegrationTest
             Editor.KeyboardCommands.Enter();
 
             Editor.KeyboardCommands.Type("\"library\":");
-            LibmanTestsUtility.WaitForCompletionEntries(Editor, expectedCompletionEntries, caseInsensitive: true, timeout: 5000);
+            CompletionTestsHelper.WaitForCompletionEntries(Editor, expectedCompletionEntries, caseInsensitive: true, timeout: 5000);
         }
     }
 }
